Rename files, simplify AutoWidthSelect
